+++ addons/yat/src/commands/Extensible.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Text;
+using YAT.Attributes;
 using YAT.Helpers;
 
 namespace YAT.Commands